+++ resources/views/layouts/main-header/navbar-right-menu/menu.blade.php
@@ -1,29 +1,29 @@
-<div class="d-flex order-lg-2 ml-auto">
+<div class="d-flex order-lg-2 ms-auto">
     <div class="nav-item d-none d-md-flex">
         <a href="https://github.com/tabler/tabler" class="btn btn-sm btn-outline-primary" target="_blank">Source code</a>
     </div>
     <div class="dropdown d-none d-md-flex">
-        <a class="nav-link icon" data-toggle="dropdown">
+        <a class="nav-link icon" href="#" data-bs-toggle="dropdown">
             <i class="fe fe-bell"></i>
             <span class="nav-unread"></span>
         </a>
-        <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
+        <div class="dropdown-menu dropdown-menu-end dropdown-menu-arrow">
             <a href="#" class="dropdown-item d-flex">
-                <span class="avatar mr-3 align-self-center" style=""></span>
+                <span class="avatar me-3 align-self-center" style=""></span>
                 <div>
                     <strong>Nathan</strong> pushed new commit: Fix page load performance issue.
                     <div class="small text-muted">10 minutes ago</div>
                 </div>
             </a>
             <a href="#" class="dropdown-item d-flex">
-                <span class="avatar mr-3 align-self-center" style=""></span>
+                <span class="avatar me-3 align-self-center" style=""></span>
                 <div>
                     <strong>Alice</strong> started new task: Tabler UI design.
                     <div class="small text-muted">1 hour ago</div>
                 </div>
             </a>
             <a href="#" class="dropdown-item d-flex">
-                <span class="avatar mr-3 align-self-center" style=""></span>
+                <span class="avatar me-3 align-self-center" style=""></span>
                 <div>
                     <strong>Rose</strong> deployed new version of NodeJS REST Api V3
                     <div class="small text-muted">2 hours ago</div>
@@ -34,14 +34,14 @@
         </div>
     </div>
     <div class="dropdown">
-        <a href="#" class="nav-link pr-0 leading-none" data-toggle="dropdown">
+        <a href="#" class="nav-link pe-0 leading-none" data-bs-toggle="dropdown">
             <span class="avatar" style=""></span>
-            <span class="ml-2 d-none d-lg-block">
+            <span class="ms-2 d-none d-lg-block">
                       <span class="text-default">Jane Pearson</span>
                       <small class="text-muted d-block mt-1">Administrator</small>
                     </span>
         </a>
-        <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
+        <div class="dropdown-menu dropdown-menu-end dropdown-menu-arrow">
             <a class="dropdown-item" href="#">
                 <i class="dropdown-icon fe fe-user"></i> Profile
             </a>
@@ -49,7 +49,7 @@
                 <i class="dropdown-icon fe fe-settings"></i> Settings
             </a>
             <a class="dropdown-item" href="#">
-                <span class="float-right"><span class="badge badge-primary">6</span></span>
+                <span class="float-end"><span class="badge bg-primary">6</span></span>
                 <i class="dropdown-icon fe fe-mail"></i> Inbox
             </a>
             <a class="dropdown-item" href="#">
@@ -78,6 +78,6 @@
         </div>
     </div>
 </div>
-<a href="#" class="header-toggler d-lg-none ml-3 ml-lg-0" data-toggle="collapse" data-target="#headerMenuCollapse">
+<a href="#" class="header-toggler d-lg-none ms-3 ms-lg-0" data-bs-toggle="collapse" data-bs-target="#headerMenuCollapse">
     <span class="header-toggler-icon"></span>
 </a>